backend/requirements.txt
flask
flask-cors
qrcode[pil]
Pillow
mysql-connector-python